Add generic CounterCommand Remove AddWinCommand Remove RemoveWinCommand
